clean up 1.0
